Add basic Event handling
This adds `Event` production to socket state. This is done via a mpsc
channel that can be consumed via `SocketState::events` which will
return a `Vec<Event>` to be consumed up the stack. This setup will
allow the `UdpSocket` to return this `Vec<Event>` back up to the user
to handle when the time is right in their implementation.

Other notable changes:
- Simple rustfmt on socket_state, udp, and events.
- Moved `Connected`, `Disconnected` and `TimedOut` to be a single
tuple of `Arc<RwLock<Connection>>` for clairty.
- Renamed `SocketState::create_connection_if_exists` to
`SocketState::get_connection_or_insert` also improved the way we use
locks in this. Since most calls to this function just needs a read
lock and not a write lock. A read lock is aquired initally to check if
there is a connection or not. If there is not, it will drop the read
lock and aquire a write lock to insert the new connection. At this
point it will also create a `Connected` event.