socket: Mitigate DOS by only adding sent-to connections
This patch fixes issue 187 by not adding incoming packet addresses to
the connection table.

A connection is now only added when explicitly sending to that
connection. This allows the user to use arbitrary code to inspect an
inbound packet and validate it by sending a response, or deciding to
ignore it.