socket: Allow manual polling (#191)
* socket: Allow manual polling

The problem with the current polling loop is that it does not wait,
causing CPU usage to skyrocket for no good reason. In a game we would
like to run `poll` during every game tick, as there is no point in
running any faster than that.

* socket: Add test for manual polling

I'm not sure if we need a synchronization primitive here but I assume
since it's on the same system that this is goes over IPC which ensures
that the server's read will always see it.

Tried with a while cargo test loop and none failed.